snapcode-backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
openai
python-dotenv
huggingface_hub
pillow
